requirements.txt
pandas
pytorch
torchvision
plotly
gradio
huggingface_hub